use quieries
